2711: Remove NodeTree interface, move docs to AABBTree
